Make Zipkin trace query available (#6454)
