+++ oap-server/server-receiver-plugin/zipkin-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/zipkin/ZipkinReceiverProvider.java
@@ -18,7 +18,6 @@
 
 package org.apache.skywalking.oap.server.receiver.zipkin;
 
-import org.apache.skywalking.oap.server.analyzer.provider.trace.parser.ISegmentParserService;
 import org.apache.skywalking.oap.server.core.CoreModule;
 import org.apache.skywalking.oap.server.library.module.ModuleConfig;
 import org.apache.skywalking.oap.server.library.module.ModuleDefine;
@@ -28,9 +27,6 @@ import org.apache.skywalking.oap.server.library.module.ServiceNotProvidedExcepti
 import org.apache.skywalking.oap.server.library.server.ServerException;
 import org.apache.skywalking.oap.server.library.server.jetty.JettyServer;
 import org.apache.skywalking.oap.server.library.server.jetty.JettyServerConfig;
-import org.apache.skywalking.oap.server.receiver.trace.module.TraceModule;
-import org.apache.skywalking.oap.server.receiver.zipkin.analysis.Receiver2AnalysisBridge;
-import org.apache.skywalking.oap.server.receiver.zipkin.analysis.transform.Zipkin2SkyWalkingTransfer;
 import org.apache.skywalking.oap.server.receiver.zipkin.handler.SpanV1JettyHandler;
 import org.apache.skywalking.oap.server.receiver.zipkin.handler.SpanV2JettyHandler;
 
@@ -82,14 +78,6 @@ public class ZipkinReceiverProvider extends ModuleProvider {
 
         jettyServer.addHandler(new SpanV1JettyHandler(config, getManager()));
         jettyServer.addHandler(new SpanV2JettyHandler(config, getManager()));
-
-        if (config.isNeedAnalysis()) {
-            ISegmentParserService segmentParseService = getManager().find(TraceModule.NAME)
-                                                                    .provider()
-                                                                    .getService(ISegmentParserService.class);
-            Receiver2AnalysisBridge bridge = new Receiver2AnalysisBridge(segmentParseService);
-            Zipkin2SkyWalkingTransfer.INSTANCE.addListener(bridge);
-        }
     }
 
     @Override
@@ -103,13 +91,6 @@ public class ZipkinReceiverProvider extends ModuleProvider {
 
     @Override
     public String[] requiredModules() {
-        if (config.isNeedAnalysis()) {
-            return new String[] {TraceModule.NAME};
-        } else {
-            /**
-             * In pure trace status, we don't need the trace receiver.
-             */
-            return new String[] {CoreModule.NAME};
-        }
+        return new String[] {CoreModule.NAME};
     }
 }
