+++ oap-server/server-receiver-plugin/zipkin-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/zipkin/trace/SpanForward.java
@@ -19,26 +19,25 @@
 package org.apache.skywalking.oap.server.receiver.zipkin.trace;
 
 import java.util.List;
+import lombok.RequiredArgsConstructor;
 import org.apache.skywalking.apm.util.StringUtil;
 import org.apache.skywalking.oap.server.core.analysis.IDManager;
-import org.apache.skywalking.oap.server.core.analysis.TimeBucket;
 import org.apache.skywalking.oap.server.core.analysis.NodeType;
+import org.apache.skywalking.oap.server.core.analysis.TimeBucket;
+import org.apache.skywalking.oap.server.core.config.NamingControl;
+import org.apache.skywalking.oap.server.core.source.EndpointMeta;
+import org.apache.skywalking.oap.server.core.source.ServiceMeta;
 import org.apache.skywalking.oap.server.core.source.SourceReceiver;
 import org.apache.skywalking.oap.server.library.util.BooleanUtils;
-import org.apache.skywalking.oap.server.receiver.zipkin.ZipkinReceiverConfig;
 import org.apache.skywalking.oap.server.receiver.zipkin.handler.SpanEncode;
 import org.apache.skywalking.oap.server.storage.plugin.zipkin.ZipkinSpan;
 import zipkin2.Span;
 import zipkin2.codec.SpanBytesEncoder;
 
+@RequiredArgsConstructor
 public class SpanForward {
-    private ZipkinReceiverConfig config;
-    private SourceReceiver receiver;
-
-    public SpanForward(ZipkinReceiverConfig config, SourceReceiver receiver) {
-        this.config = config;
-        this.receiver = receiver;
-    }
+    private final NamingControl namingControl;
+    private final SourceReceiver receiver;
 
     public void send(List<Span> spanList) {
         spanList.forEach(span -> {
@@ -49,27 +48,28 @@ public class SpanForward {
             if (StringUtil.isEmpty(serviceName)) {
                 serviceName = "Unknown";
             }
+            serviceName = namingControl.formatServiceName(serviceName);
             zipkinSpan.setServiceId(IDManager.ServiceID.buildId(serviceName, NodeType.Normal));
 
-            String spanName = span.name();
-            Span.Kind kind = span.kind();
-            switch (kind) {
-                case SERVER:
-                case CONSUMER:
-                    if (!StringUtil.isEmpty(spanName)) {
-                        zipkinSpan.setEndpointId(IDManager.EndpointID.buildId(zipkinSpan.getServiceId(), span.name()));
-                    }
-            }
-            if (!StringUtil.isEmpty(spanName)) {
-                zipkinSpan.setEndpointName(spanName);
-            }
             long startTime = span.timestampAsLong() / 1000;
             zipkinSpan.setStartTime(startTime);
-            if (startTime != 0) {
-                long timeBucket = TimeBucket.getRecordTimeBucket(zipkinSpan.getStartTime());
-                zipkinSpan.setTimeBucket(timeBucket);
-            }
+            long timeBucket = TimeBucket.getRecordTimeBucket(zipkinSpan.getStartTime());
+            zipkinSpan.setTimeBucket(timeBucket);
 
+            String spanName = span.name();
+            if (!StringUtil.isEmpty(spanName)) {
+                final String endpointName = namingControl.formatEndpointName(serviceName, spanName);
+                zipkinSpan.setEndpointName(endpointName);
+                zipkinSpan.setEndpointId(IDManager.EndpointID.buildId(zipkinSpan.getServiceId(), endpointName));
+
+                //Create endpoint meta for the server side span
+                EndpointMeta endpointMeta = new EndpointMeta();
+                endpointMeta.setServiceName(serviceName);
+                endpointMeta.setServiceNodeType(NodeType.Normal);
+                endpointMeta.setEndpoint(endpointName);
+                endpointMeta.setTimeBucket(timeBucket);
+                receiver.receive(endpointMeta);
+            }
             long latency = span.durationAsLong() / 1000;
 
             zipkinSpan.setEndTime(startTime + latency);
@@ -78,7 +78,19 @@ public class SpanForward {
             zipkinSpan.setLatency((int) latency);
             zipkinSpan.setDataBinary(SpanBytesEncoder.PROTO3.encode(span));
 
+            span.tags().forEach((key, value) -> {
+                zipkinSpan.getTags().add(key + "=" + value);
+            });
+
             receiver.receive(zipkinSpan);
+
+            // Create the metadata source
+            // No instance name is required in the Zipkin model.
+            ServiceMeta serviceMeta = new ServiceMeta();
+            serviceMeta.setName(serviceName);
+            serviceMeta.setNodeType(NodeType.Normal);
+            serviceMeta.setTimeBucket(timeBucket);
+            receiver.receive(serviceMeta);
         });
     }
 }
