+++ tools/dependencies/known-oap-backend-dependencies.txt
@@ -11,7 +11,6 @@ bcpkix-jdk15on-1.66.jar
 bcprov-ext-jdk15on-1.66.jar
 bcprov-jdk15on-1.66.jar
 builder-annotations-0.22.0.jar
-caffeine-2.6.2.jar
 checker-qual-2.8.1.jar
 client-java-10.0.0.jar
 client-java-api-10.0.0.jar
@@ -163,7 +162,6 @@ swagger-annotations-1.6.2.jar
 t-digest-3.2.jar
 vavr-0.10.3.jar
 vavr-match-0.10.3.jar
-zipkin-2.9.1.jar
 zookeeper-3.4.10.jar
 kafka-clients-2.4.1.jar
 lz4-java-1.6.0.jar
